Update course server ( info model )
1. info model의 소수를 사용하는 컬럼에 값의 범위가 커서 DOUBLE로 변경

+++ Backend/Course/models/info.js
@@ -37,18 +37,18 @@ module.exports = (sequelize, DataTypes) => {
         },
     
         grade_avg: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.DOUBLE,
             allowNULL: true,
             defaultValue: 0
         },
 
         latitude: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DOUBLE,
             allowNULL: true
         },
         
         longitude: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DOUBLE,
             allowNULL: true
         },
